land-classifier/requirements.txt
Flask==1.1.4
python-dotenv==0.15.0
flask_cors==3.0.10
functions-framework==2.1.0
google-cloud-storage==1.41.1
google-api-python-client==2.20.0
google-cloud-core==1.7.2
google-cloud-monitoring
google-cloud-logging==2.6.0
